Add model and simple classification code
Add in the model so that we can add it to the Docker image (could have
probably kept it out of git actually) as well as the ImageNet
classifications in JSON format.

Add in a sample picture.

Add logic to the Flask app to return the predictions rather than a
string.

Issue(s): None